Add promote_strict mechanism and use it instead of typejoin() where appropriate
Introduce a mechanism similar to promote(), but used to return types which can hold
exactly all values of the input types. Use it instead of typejoin() to choose an
appropriate element type with collect(), map(), broadcast(), and Dict().

Add appropriate methods for Number types, as well as for Nothing/Missing so that
Union{T, Nothing/Missing} is used instead of Any. Add promotion rules for Tuple and
NamedTuple so that promotion is performed element by element, to allow for more
precise typing of fields, potentially helping the compiler down the road.